### 2019-11-08 #1877 @whedon
[PRE REVIEW]: rnassqs: An R package to access agricultural data via the USDA National Agricultural Statistical Service (USDA-NASS) 'Quick Stats' API
**Submitting author:** @potterzot (<a href="http://orcid.org/0000-0002-3410-3732">Nicholas A Potter</a>)
**Repository:** <a href="https://github.com/ropensci/rnassqs" target ="_blank">https://github.com/ropensci/rnassqs</a>
**Version:** v0.6.1
