### 2019-11-08 #1877 @kyleniemeyer
Hi @potterzot, it looks like you are missing a `paper.bib` file in the repo—do you have any references?
