### 2019-11-08 77f5a2b
Remove reference to paper.bib since no references
