### 2019-11-08 88bfbd1
Squashing the following commits:
- add paper.bib since seems required for JOSS compile.
- add zenodo metadata.
- change version number.
- regen docs and codemeta.
