### 2019-11-08 #1877 @potterzot
@kyleniemeyer ack, sorry! I removed it because when cleaning up the paper for submission I ended up without any references. I thought I had retested the compile but apparently not. I've added an empty paper.bib so that it compiles correctly.
