### 2019-11-08 #1877 @potterzot
For reference, I'm submitting as part of the rOpenSci review, which I believe I am supposed to post here: https://github.com/ropensci/software-review/issues/298.
