### 2019-11-09 #1880 @kyleniemeyer
For the record, this is an rOpenSci-accepted submission; the review thread is here: https://github.com/ropensci/software-review/issues/298
