### 2019-11-09 #1880 @kyleniemeyer
Hi @potterzot, I was about to add the Zenodo DOI here, but I noticed that the metadata doesn't look quite right—can you make the title and authors at 10.5281/zenodo.3532167 match that of the paper?
