### 2019-11-12 #1880 @potterzot
@kyleniemeyer ah I see the problem. I have contributors and reviewers from the rOpenSci review process listed in DESCRIPTION, which was being converted to the zenodo metadata. I've added a '.zenodo.json' file to explicitly define authorship. Contributors and reviewers are thanked in the acknowledgement section of the paper.

Should I create a new zenodo DOI at this point as well?
