### 2019-11-12 #1880 @kyleniemeyer
@potterzot 👍

If nothing in the software changed, you can just modify the authors and title for the existing version (in Zenodo you can alter the metadata for something published, though you can't modify the files). If anything in the repo changed, other than that `.zenodo.json` file, then you should create a new version and let me know the DOI.
