### 2019-11-12 #1880 @potterzot
@kyleniemeyer thank you, I didn't realize that was possible. I modified the zenodo entry. Nothing else has changed so I did not create a new DOI.
